error x10 en densidad solucionado

+++ sim/grid.hpp
@@ -70,7 +70,7 @@ class Grid {
     std::vector<double> accelerationY;
     std::vector<double> accelerationZ;
     std::vector<double> density;
-    float ppm;
+    double ppm;
     int num_particles;
     double h;
     double particle_mass;
@@ -93,7 +93,6 @@ class Grid {
     int readFile(std::string const & input_file_name);
     int writeFile(std::string const & output_file_name);
 
-    
     void generateParticlePairs();
     void positionateParticle();
     void resetBlocks();
@@ -107,7 +106,6 @@ class Grid {
     void interactionsXGrid();
     void interactionsYGrid();
     void interactionsZGrid();
-    
 
   private:
     bool readHeader(std::ifstream & input_file);
